refactor: 기물 클래스 불변 객체 선언
