+++ src/main/java/domain/piece/Elephant.java
@@ -1,14 +1,14 @@
 package domain.piece;
 
-import domain.Direction;
+import domain.coordinate.Direction;
 import domain.Game;
-import domain.Position;
+import domain.coordinate.Position;
 import domain.Side;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Elephant extends Piece {
+public final class Elephant extends Piece {
 
     public Elephant(Side side) {
         super(side);
